requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiomysql>=0.2.0
mysql-connector-python>=8.0.0
phpserialize>=1.3
anthropic>=0.25.0
requests>=2.31.0
